+++ python-sdk/examples/openai_bot_prompt.py
@@ -6,7 +6,7 @@ from openai import OpenAI
 
 client = OpenAI()
 
-import langwatch.prompts
+import langwatch.prompt
 import asyncio
 
 langwatch.setup()
@@ -23,7 +23,7 @@ async def main():
     # Raw prompt {model: 'gpt-4o-mini', messages: [{ role: 'system', content: 'The user is {user_name} and their email is {user_email}' }]}
     # Autobuilt prompt { model: 'gpt-4o-mini', messages: [{ role: 'system', content: 'The user is John Doe and their email is john.doe@example.com' }]}
     prompt_id = "prompt_a6EDs9WsgnZuq9KnVbOXc"
-    prompt = langwatch.prompts.get_prompt(prompt_id)
+    prompt = langwatch.prompt.get_prompt(prompt_id)
     print(prompt.raw_config())
     messages = prompt.format_messages(input="I like to eat pizza")
     completion = client.chat.completions.create(model=prompt.model, messages=messages)
